requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.4
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.24.4
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
pydantic==2.3.0
pydantic_core==2.6.3
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.23.2
zipp==3.17.0

